Update checkOutcome to determine if bots are the same position
